requirements.txt
torch<3
numpy<3
hydra-core
tqdm
torchtext
pytorch-nlp
spacy
